Add three classes in Android Test folder. Functional TestSuite: This is a testsuite and currently both test classes are added into the suties TestApproach1 : In this class you will find 1 test case which will run the whole flow. TestApproach2: In this class you will find multiple testcase which are covering the flow.
The flow / test cases which I have covered in this assingment are as follow :

t1VerifyUserIsOnHomePage : This testcase will verify that user is on home page.
t2VerifyUserIsWatchingCorrectCityWeather : This test case will verify that user is navigated to the right city which he clicked from home page for weather detail.
t3VeifyUserNavigatedBacktoMainUI : This test case will verify that use is navigated back to home page.
t4VerifyaddNewCity : This test case will add a new city and verify that newly added city is appearing on  home page.
t5DeleteCityFromListingView : This test case will delete the newly added city. Note: Currently This test case is not working due to the variable scope issue. so I have added this operation in Testcase4 at the end